Env: Use cytomic bandpass as the pitch pre-filter.
Experimenting, I found that Gravy made a good pre-filter
for Env's pitch detector. So instead of using the simple
lo/hi-cut pair of filters for bandpass, use the same Cytomic
filter as I use in Gravy. Use bandpass mode for now, and
expose FREQ and RES control groups just like Gravy does.
FREQ helps center on the expected detected pitches, and
RES helps squeeze the frequency response near that center
frequency.

+++ src/gravy_engine.hpp
@@ -8,7 +8,7 @@ namespace Sapphire
     namespace Gravy
     {
         const int OctaveRange = 5;                              // +/- octave range around default frequency
-        const float DefaultFrequencyHz = 523.2511306011972;     // C5 = 440*(2**0.25)
+        const double DefaultFrequencyHz = 523.2511306011972;     // C5 = 440*(2**0.25)
         const int FrequencyFactor = 1 << OctaveRange;
 
         const float DefaultFrequencyKnob = 0.0;
